+++ src/cartesian/Area.js
@@ -5,7 +5,6 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import Animate from 'react-smooth';
-import { interpolateNumber } from 'd3-interpolate';
 import _ from 'lodash';
 import Curve from '../shape/Curve';
 import Dot from '../shape/Dot';
@@ -14,7 +13,7 @@ import LabelList from '../component/LabelList';
 import pureRender from '../util/PureRender';
 import { PRESENTATION_ATTRIBUTES, EVENT_ATTRIBUTES, LEGEND_TYPES,
   getPresentationAttributes, isSsr } from '../util/ReactUtils';
-import { isNumber, uniqueId } from '../util/DataUtils';
+import { isNumber, uniqueId, interpolateNumber } from '../util/DataUtils';
 import { getCateCoordinateOfLine, getValueByDataKey } from '../util/ChartUtils';
 
 
@@ -313,6 +312,13 @@ class Area extends Component {
 
     return (
       <Layer clipPath={needClip ? `url(#clipPath-${this.id})` : null}>
+        <Curve
+          {...this.props}
+          points={points}
+          baseLine={baseLine}
+          stroke="none"
+          className="recharts-area-area"
+        />
         {stroke !== 'none' && (
           <Curve
             {...getPresentationAttributes(this.props)}
@@ -335,13 +341,6 @@ class Area extends Component {
             points={baseLine}
           />
         )}
-        <Curve
-          {...this.props}
-          points={points}
-          baseLine={baseLine}
-          stroke="none"
-          className="recharts-area-area"
-        />
       </Layer>
     );
   }