+++ src/cartesian/Line.js
@@ -4,7 +4,6 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import Animate from 'react-smooth';
-import { interpolateNumber } from 'd3-interpolate';
 import classNames from 'classnames';
 import _ from 'lodash';
 import pureRender from '../util/PureRender';
@@ -13,7 +12,7 @@ import Dot from '../shape/Dot';
 import Layer from '../container/Layer';
 import LabelList from '../component/LabelList';
 import ErrorBar from './ErrorBar';
-import { uniqueId } from '../util/DataUtils';
+import { uniqueId, interpolateNumber } from '../util/DataUtils';
 import { PRESENTATION_ATTRIBUTES, EVENT_ATTRIBUTES, LEGEND_TYPES, filterEventAttributes,
   getPresentationAttributes, isSsr, findChildByType } from '../util/ReactUtils';
 import { getCateCoordinateOfLine, getValueByDataKey } from '../util/ChartUtils';
@@ -282,6 +281,7 @@ class Line extends Component {
   }
 
   renderCurveStatically(points, needClip, props) {
+    const { type, layout, connectNulls } = this.props;
     const curveProps = {
       ...getPresentationAttributes(this.props),
       ...filterEventAttributes(this.props),
@@ -290,6 +290,7 @@ class Line extends Component {
       clipPath: needClip ? `url(#clipPath-${this.id})` : null,
       points,
       ...props,
+      type, layout, connectNulls,
     };
 
     return <Curve {...curveProps} pathRef={this.pathRef} />;