+++ src/polar/Radar.js
@@ -6,9 +6,9 @@ import PropTypes from 'prop-types';
 import Animate from 'react-smooth';
 import classNames from 'classnames';
 import _ from 'lodash';
-import { interpolateNumber } from 'd3-interpolate';
+import { interpolateNumber } from '../util/DataUtils';
 import pureRender from '../util/PureRender';
-import { PRESENTATION_ATTRIBUTES, LEGEND_TYPES,
+import { PRESENTATION_ATTRIBUTES, LEGEND_TYPES, filterEventAttributes,
   getPresentationAttributes, isSsr } from '../util/ReactUtils';
 import { polarToCartesian } from '../util/PolarUtils';
 import { getValueByDataKey } from '../util/ChartUtils';
@@ -180,6 +180,7 @@ class Radar extends Component {
     } else {
       radar = (
         <Polygon
+          {...filterEventAttributes(this.props)}
           onMouseEnter={this.handleMouseEnter}
           onMouseLeave={this.handleMouseLeave}
           {...getPresentationAttributes(this.props)}